+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added back `netstandard2.0` target.
+([#3787](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3787))
+
 ## 1.0.0-rc9.8
 
 Released 2022-Oct-17